Make sure the embed wrapper div is getting stripped on the project renderer as well
