+++ apps/project-renderer/package.json
@@ -26,7 +26,6 @@
     "@crowdsignal/http": "^0.1.0",
     "@crowdsignal/rest-api": "^0.1.0",
     "@crowdsignal/router": "^0.1.0",
-    "@wordpress/api-fetch": "^5.2.6",
     "@wordpress/element": "^4.0.4",
     "@wordpress/i18n": "^4.2.4",
     "lodash": "^4.17.21"
